Refactor game service & controller

+++ src/main/java/com/spaf/coderush/game/model/CreateGameRequest.java
@@ -10,5 +10,5 @@ public class CreateGameRequest {
     private List<String> languages;
     private byte rounds;
     private String password;
-    private int roundTimeLimit;
+    private short roundTimeLimit;
 }
